working on copying notes to client to see
