+++ openpype/modules/ftrack/event_handlers_user/action_ttd_copy_delivery_notes.py
@@ -1,111 +1,13 @@
 from typing import List
-from tempfile import gettempdir
-from urllib.request import urlopen
-from shutil import copyfileobj
-from os import environ
-from logging import Logger, getLogger
+from logging import getLogger
 
 logger = getLogger(__name__)
 
-from openpype_modules.ftrack.lib import BaseAction, statics_icon # type: ignore
-from ftrack_api import Session, symbol
+from openpype_modules.ftrack.lib import BaseAction, statics_icon, create_notes # type: ignore
+from ftrack_api import Session
 from ftrack_api.entity.base import Entity
 from ftrack_api.entity.asset_version import AssetVersion
-from ftrack_api.entity.component import Component
-from ftrack_api.entity.note import Note
 from ftrack_api.event.base import Event
-from ftrack_api.entity.location import Location
-
-matching_fields = [
-    "content",
-    "project_id",
-    "parent_type",
-    "category_id",
-    "user_id",
-    ]
-
-def duplicate_ftrack_server_component(comp: Component, session: Session):
-
-    # server_location: Location = Session().get('Location', symbol.SERVER_LOCATION_ID)
-
-    for cloc in comp["component_locations"]:
-
-        if cloc["location"]["name"] == "ftrack.server":
-            server_location: Location = cloc['location']
-            print(f"Fetching component from location {cloc['location']}")
-            try:
-                url = server_location.get_url(comp)
-            except Exception as e:
-                print(f"Failed to retrieve url for {comp} in {server_location} due to {e}")
-                raise (e)
-            name = comp["name"] + comp["file_type"]
-            f = gettempdir() + "/" + name
-            print(f"Creating component from temp file: {f}")
-
-            with urlopen(url) as response, open(f, 'wb') as out_file:
-                copyfileobj(response, out_file)
-        else:
-            print(f"Ignoring location {cloc['location']['name']}")
-            # it is unmanaged, maybe the source is locally and thus inaccesible
-            continue
-        return session.create_component(f, {"name":name}, location=cloc["location"])
-
-
-def transfer_note_components(note_src: Note, note_target: Note, session: Session):
-    for note_comp in note_src["note_components"]:
-        print(f"Working on component {note_comp}")
-        comp = note_comp["component"]
-        if any(comp["name"] != c["component"] for c in note_target["note_components"]):
-            print(f"This component was already copied {comp['name']}")
-            continue
-        try:
-            new_component = duplicate_ftrack_server_component(comp, session)
-        except Exception as e:
-            print(f"Failed to copy component {comp} due to {e}")
-            continue
-        if new_component is None:
-            continue
-        session.create("NoteComponent", {"component_id":new_component["id"], "note_id": note_target["id"]})
-
-
-def create_notes(
-    session: Session, src_note: Note, target: Note, notes_in_target: List[Note]
-    ):
-    msg = ""
-    # REMOVE ALREADY MATCHING NOTES IN TARGET
-    for target_note in notes_in_target:
-        if  all(src_note[f] == target_note[f] for f in matching_fields):
-
-            logger.info(
-                f"Note already copied. {src_note} -> {target_note}, "
-                f"removing it. Content is: {src_note['content']}"
-            )
-            session.delete(target_note)
-    else:
-        replies = [
-            session.create(
-                "Note",
-                {tag: reply[tag] for tag in matching_fields}
-                ) for reply in src_note["replies"]
-            ]
-
-        logger.info("Note components are:", list(src_note["note_components"]))
-
-        r = session.create("Note", {
-            **{tag: src_note[tag] for tag in matching_fields},
-            "parent_id": target["id"],
-            "replies": replies,
-            # "note_label_links" : note_label_links,
-            })
-        
-        transfer_note_components(src_note, r, session)
-
-        for i, reply in enumerate(src_note["replies"]):
-            transfer_note_components(reply, r["replies"][i], session)
-
-        logger.info(f"Updated client notes for version_id {target['id']}")
-        msg += f"Updated client notes for version {target['id']}"
-    return msg
 
 class CopyDeliveryNotes(BaseAction):
     """Action for forwarding client notes from delivery version into source."""
@@ -127,8 +29,6 @@ class CopyDeliveryNotes(BaseAction):
         if is_valid:
             is_valid = self.valid_roles(session, entities, event)
         return is_valid
-
-
 
     def copy_client_notes(self, session: Session, versions: List[AssetVersion]):
